Move `GestureHandlerRootView` to `components` directory
