+++ src/components/GestureHandlerRootView.android.tsx
@@ -1,9 +1,9 @@
 import * as React from 'react';
 import { PropsWithChildren } from 'react';
 import { ViewProps } from 'react-native';
-import { maybeInitializeFabric } from './init';
-import GestureHandlerRootViewContext from './GestureHandlerRootViewContext';
-import GestureHandlerRootViewNativeComponent from './specs/RNGestureHandlerRootViewNativeComponent';
+import { maybeInitializeFabric } from '../init';
+import GestureHandlerRootViewContext from '../GestureHandlerRootViewContext';
+import GestureHandlerRootViewNativeComponent from '../specs/RNGestureHandlerRootViewNativeComponent';
 
 export interface GestureHandlerRootViewProps
   extends PropsWithChildren<ViewProps> {}
