+++ src/components/GestureHandlerRootView.tsx
@@ -1,8 +1,8 @@
 import * as React from 'react';
 import { PropsWithChildren } from 'react';
 import { View, ViewProps } from 'react-native';
-import { maybeInitializeFabric } from './init';
-import GestureHandlerRootViewContext from './GestureHandlerRootViewContext';
+import { maybeInitializeFabric } from '../init';
+import GestureHandlerRootViewContext from '../GestureHandlerRootViewContext';
 
 export interface GestureHandlerRootViewProps
   extends PropsWithChildren<ViewProps> {}
